requirements.txt
# ─── Core framework ───────────────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langsmith>=0.1.0

# ─── LLM providers (all optional — configure via LLM_PROVIDER env var) ────────
langchain-google-genai>=2.0.0   # Gemini (default/free-tier)
langchain-openai>=0.2.0          # OpenAI (optional)
# langchain-anthropic>=0.3.0    # Anthropic (optional, uncomment if needed)

# ─── Embeddings ───────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0    # local BAAI/bge-small-en-v1.5

# ─── Vector store ─────────────────────────────────────────────────────────────
qdrant-client>=1.9.0            # in-memory or server mode

# ─── Keyword / BM25 ──────────────────────────────────────────────────────────
rank-bm25>=0.2.2

# ─── Document parsing ─────────────────────────────────────────────────────────
pymupdf>=1.24.0                 # PyMuPDF for PDF → markdown
python-docx>=1.1.0              # DOCX parsing

# ─── API ──────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
httpx>=0.27.0
